799 Move action serverRelativeUrl to model, use buildUrl
Removes the ulurNumberLink component in favor of housing
serverRelativeUrl logic in the Action model and buildUrl helper.
