--- client/app/templates/components/show-project/actions-list/ulurp-number-link.hbs
@@ -1,29 +1,0 @@
-{{#if @action.dcpUlurpnumber}}
-  ULURP Number:
-  {{#if (in-array
-      (array "Approved" "Certified" "Referred" "Disapproved" "Terminated" "Withdrawn" "Study Complete")
-      @action.statuscode
-    )
-  }}  
-      {{#if (and (not-eq @action.statuscode this.actionStatuscodeActive) @action.dcpSpabsoluteurl)}}
-        {{#tool-tipster
-          content='City Planning Commision Report'
-          tagName='span'
-        }}
-          <a
-            href={{concat this.host '/document/projectaction' this.serverRelativeUrl}}
-            target="_blank"
-          >
-            {{@action.dcpUlurpnumber}}
-            {{fa-icon 'external-link-alt'}}
-          </a>
-        {{/tool-tipster}}
-      {{else}}
-        {{@action.dcpUlurpnumber}}
-      {{/if}}
-  {{else}}
-    {{@action.dcpUlurpnumber}}
-  {{/if}}
-{{else}}
-  No ULURP Number
-{{/if}}
